Extracted the shared merge functionality to it's own module where a mixin is created based on identity and factory passed to it
